move review from feature to cogs, some fixes for banned roles and anonym
